requirements.txt
elasticsearch>=8.0.0
sentence-transformers
Pillow
transformers
torch
dotenv
python-dotenv